1402-1403 second term/cloud/HW1/requirements.txt
flask
requests
pika
boto3
python-dotenv
psycopg2